Astrotrac 360 On Coord Set Slew
When the ON_COORD_SET property is set to "Slew" in the main control tab,
the mount currently slews to the target coordinate and automatically
starts tracking (as defined for the behviour of "Track"). With the new
behaviour, the mount slews to the target coordinate and stops acccording
to the definition of "Slew" described at https://www.indilib.org/
developers/deveioper-manual/101-standard-properties.html#h3-
telescopes.Tested on the AT360 mount.